quickstart by platform
One takeaway from TW7 is we want better windows support. With three
platforms it becomes increasingly hard to have one guide that works
everywhere. This bites the bullet and creates a page for each platform
where we can put all of the instructions without having readers skip
around throughout a long document.

Unfortunately I don't know how to run Tock on Windows, so that guide is
a little sparse. We need to update that, but now we have a clear place
to do it.